feat: support decimal BHK with 0.5 increments (.5, 1.5, 2.5, 3.5) across backend and frontend

+++ backend/src/main/java/com/maplistings/config/DataSeeder.java
@@ -230,7 +230,7 @@ public class DataSeeder implements CommandLineRunner {
         log.info("Successfully seeded {} listings (Residential & Commercial).", listings.size());
     }
 
-    private Listing createListing(String title, BigDecimal price, int bhk, int areaSqFt,
+    private Listing createListing(String title, BigDecimal price, double bhk, int areaSqFt,
                                    PropertyCategory propertyCategory, PropertyType propertyType,
                                    ListingType listingType,
                                    String projectName, String societyName, String address,
